multigpu fixes | regression fixes | better image detection

+++ helpers/data_backend/factory.py
@@ -15,7 +15,6 @@ from helpers.training.state_tracker import StateTracker
 import json, os, torch, logging, io
 
 logger = logging.getLogger("DataBackendFactory")
-logger.setLevel(os.environ.get("SIMPLETUNER_LOG_LEVEL", "INFO"))
 
 
 def init_backend_config(backend: dict, args: dict, accelerator) -> dict:
@@ -206,6 +205,11 @@ def configure_multi_databackend(
     """
     Configure a multiple dataloaders based on the provided commandline args.
     """
+    logger.setLevel(
+        os.environ.get(
+            "SIMPLETUNER_LOG_LEVEL", "INFO" if accelerator.is_main_process else "ERROR"
+        )
+    )
     if args.data_backend_config is None:
         raise ValueError(
             "Must provide a data backend config file via --data_backend_config"
@@ -270,7 +274,7 @@ def configure_multi_databackend(
             raise ValueError(f"Unknown data backend type: {backend['type']}")
 
         preserve_data_backend_cache = backend.get("preserve_data_backend_cache", False)
-        if not preserve_data_backend_cache:
+        if not preserve_data_backend_cache and accelerator.is_local_main_process:
             StateTracker.delete_cache_files(
                 data_backend_id=init_backend["id"],
                 preserve_data_backend_cache=preserve_data_backend_cache,
@@ -298,7 +302,7 @@ def configure_multi_databackend(
                 and args.caption_dropout_probability > 0
             ):
                 logger.info("Pre-computing null embedding for caption dropout")
-                with accelerator.main_process_first():
+                if accelerator.is_local_main_process:
                     init_backend["text_embed_cache"].compute_embeddings_for_prompts(
                         [""], return_concat=False, load_from_cache=False
                     )
@@ -412,29 +416,32 @@ def configure_multi_databackend(
                 )
         else:
             raise ValueError(f"Unknown metadata backend type: {metadata_backend}")
-        init_backend["metadata_backend"] = BucketManager_cls(
-            id=init_backend["id"],
-            instance_data_root=init_backend["instance_data_root"],
-            data_backend=init_backend["data_backend"],
-            accelerator=accelerator,
-            resolution=backend.get("resolution", args.resolution),
-            minimum_image_size=backend.get(
-                "minimum_image_size", args.minimum_image_size
-            ),
-            resolution_type=backend.get("resolution_type", args.resolution_type),
-            batch_size=args.train_batch_size,
-            metadata_update_interval=backend.get(
-                "metadata_update_interval", args.metadata_update_interval
-            ),
-            cache_file=os.path.join(
-                init_backend["instance_data_root"], "aspect_ratio_bucket_indices.json"
-            ),
-            metadata_file=os.path.join(
-                init_backend["instance_data_root"], "aspect_ratio_bucket_metadata.json"
-            ),
-            delete_problematic_images=args.delete_problematic_images or False,
-            **metadata_backend_args,
-        )
+        with accelerator.main_process_first():
+            init_backend["metadata_backend"] = BucketManager_cls(
+                id=init_backend["id"],
+                instance_data_root=init_backend["instance_data_root"],
+                data_backend=init_backend["data_backend"],
+                accelerator=accelerator,
+                resolution=backend.get("resolution", args.resolution),
+                minimum_image_size=backend.get(
+                    "minimum_image_size", args.minimum_image_size
+                ),
+                resolution_type=backend.get("resolution_type", args.resolution_type),
+                batch_size=args.train_batch_size,
+                metadata_update_interval=backend.get(
+                    "metadata_update_interval", args.metadata_update_interval
+                ),
+                cache_file=os.path.join(
+                    init_backend["instance_data_root"],
+                    "aspect_ratio_bucket_indices.json",
+                ),
+                metadata_file=os.path.join(
+                    init_backend["instance_data_root"],
+                    "aspect_ratio_bucket_metadata.json",
+                ),
+                delete_problematic_images=args.delete_problematic_images or False,
+                **metadata_backend_args,
+            )
 
         if "aspect" not in args.skip_file_discovery or "aspect" not in backend.get(
             "skip_file_discovery", ""
@@ -460,6 +467,7 @@ def configure_multi_databackend(
         init_backend["metadata_backend"].split_buckets_between_processes(
             gradient_accumulation_steps=args.gradient_accumulation_steps,
         )
+        accelerator.wait_for_everyone()
 
         # Check if there is an existing 'config' in the metadata_backend.config
         excluded_keys = [
@@ -559,7 +567,7 @@ def configure_multi_databackend(
                 f"Backend {init_backend['id']} has prepend_instance_prompt=True, but no instance_prompt was provided. You must provide an instance_prompt, or disable this option."
             )
 
-        with accelerator.main_process_first():
+        if accelerator.is_local_main_process:
             # We get captions from the IMAGE dataset. Not the text embeds dataset.
             captions = PromptHandler.get_all_captions(
                 data_backend=init_backend["data_backend"],
@@ -569,6 +577,7 @@ def configure_multi_databackend(
                 use_captions=use_captions,
                 caption_strategy=backend.get("caption_strategy", args.caption_strategy),
             )
+
             if "text" not in args.skip_file_discovery and "text" not in backend.get(
                 "skip_file_discovery", ""
             ):
@@ -619,10 +628,11 @@ def configure_multi_databackend(
             vae_cache_preprocess=args.vae_cache_preprocess,
         )
 
-        logger.info(f"(id={init_backend['id']}) Discovering cache objects..")
-        if accelerator.is_local_main_process:
-            init_backend["vaecache"].discover_all_files()
-        accelerator.wait_for_everyone()
+        if args.vae_cache_preprocess:
+            logger.info(f"(id={init_backend['id']}) Discovering cache objects..")
+            if accelerator.is_local_main_process:
+                init_backend["vaecache"].discover_all_files()
+            accelerator.wait_for_everyone()
 
         if (
             (
@@ -651,8 +661,10 @@ def configure_multi_databackend(
             init_backend["metadata_backend"].load_image_metadata()
         accelerator.wait_for_everyone()
 
-        if "vae" not in args.skip_file_discovery and "vae" not in backend.get(
-            "skip_file_discovery", ""
+        if (
+            args.vae_cache_preprocess
+            and "vae" not in args.skip_file_discovery
+            and "vae" not in backend.get("skip_file_discovery", "")
         ):
             init_backend["vaecache"].split_cache_between_processes()
             if args.vae_cache_preprocess:
